rename: chat -> chatbot 도메인명 변경
